cods refactoring
