+++ client/src/components/channelOption.tsx
@@ -1,4 +1,5 @@
 import Dropdown from 'react-bootstrap/Dropdown';
+import { useTranslation } from 'react-i18next';
 
 interface IProps {
   id: string;
@@ -7,14 +8,20 @@ interface IProps {
 }
 const ChannelOption = (props: IProps) => {
   const { id, onDelete, onEdit } = props;
+  const { t } = useTranslation();
+
   return (
     <Dropdown>
       <Dropdown.Toggle variant='Secondary' id={id}>
-        <span className='visually-hidden'>Управление каналом</span>
+        <span className='visually-hidden'>{t('channelMenu.control')}</span>
       </Dropdown.Toggle>
       <Dropdown.Menu>
-        <Dropdown.Item onClick={() => onDelete()}>Удалить</Dropdown.Item>
-        <Dropdown.Item onClick={() => onEdit()}>Переименовать</Dropdown.Item>
+        <Dropdown.Item onClick={() => onDelete()}>
+          {t('channelMenu.delete')}
+        </Dropdown.Item>
+        <Dropdown.Item onClick={() => onEdit()}>
+          {t('channelMenu.rename')}
+        </Dropdown.Item>
       </Dropdown.Menu>
     </Dropdown>
   );
