+++ client/src/types/store.ts
@@ -1,27 +1,33 @@
-export type Channels = {
-    ids: string[],
-    entities: {
-        [key: string]: Channel
-    },
-    currentChannelId: string,
-}
+export type IChannels = {
+  ids: string[];
+  entities: {
+    [key: string]: Channel;
+  };
+  currentChannelId: string;
+};
 
 export type Channel = {
-    id: string,
-    name: string,
-    removable: boolean,
-}
+  id: string;
+  name: string;
+  removable: boolean;
+};
 
 export type Messages = {
-    ids: string[],
-    entities: {
-        [key: string]: Message
-    },
-}
+  ids: string[];
+  entities: {
+    [key: string]: Message;
+  };
+};
 
 export type Message = {
-    id: string,
-    channelId: string,
-    name: string,
-    text: string,
-}
+  id: string;
+  channelId: string;
+  name: string;
+  text: string;
+};
+
+export type TWordMessage = {
+  manyCase: string;
+  singularCase: string;
+  genitiveCase: string;
+};
